Fixed issue #14 and #15
