+++ test/runtests.jl
@@ -1,14 +1,13 @@
-using SearchableGeometries
 using SafeTestsets
 
 @safetestset "SearchableGeometries.jl" begin
     @safetestset "BoundingVolumes:" begin
-        include("BoundingVolume_test.jl")
+        include("GeometricPrimitives/BoundingVolume_test.jl")
     end
     @safetestset "Balls:" begin
-        include("Ball_test.jl")
+        include("GeometricPrimitives/Ball_test.jl")
     end
     @safetestset "Hyperplanes:" begin
-        include("Hyperplane_test.jl")
+        include("GeometricPrimitives/Hyperplane_test.jl")
     end
 end
